Fixed an issue within the fu library (Py3-only) which resulted in a cell tuple being "occasionally" returned as being empty in Py3.
This is an incredibly stupid bug that doesn't seem to occur in Py2..However,
in my instance of Py3.9.12 the closure tuple will "sometimes" be returned
as empty...for no explainable reason..and on completely different executions of
the testcase within the same instance of the interpreter. This ended up being
fixed by storing it in a list first, and then converting it back into a tuple
prior to actually returning it back to the caller.