Demoted an unnecessary exception to a warning when too much data was loaded into a `ptype.container` from the ptypes library.
Also, the entire ptypes library was modified to remove all backtraces that were
being rendered as part of the exception. Some of the parameters for the
exceptions were also tweaked to include the offset and size that caused them.

The backtrace was originally necessary so that you could figure out which type
was being decoded incorrectly. Now it should be unnecessary because you should
be able to figure that out using the offset and size (from the exception that
was raised) to navigate to the either the nearest or the last decoded field.