Modified the ptypes.parray library to ensure that all parray types are allocated incrementally.
For performance reasons, the ptypes.parray library would typically rely
on its parent class to perform a load from the empty provider during
allocation. After the array has been loaded with empty data, the
type.alloc method would then assign any allocated types into the
already-initialized array.

The introduction of commit baf77c1d5e20a035054532350ef5a58482f4d9cd
added the ability for the type.alloc method to assign elements to the
array non-sequentially (with a dictionary). That commit actually fixed
the original implementation which would detect if a tuple was passed as
an element and use it to determine the key and value to update. The
reason why that commit was created in the first place is because a
caller might implement their own "alloc()" method that takes a tuple as
its single parameter.

The combination of the specified commit and the original implementation
of the type.alloc method resulted in the offset being calculated
incorrectly. This poses an issue when the blocksize for the element
being added is based on its current offset. This commit fixes the issue
by ensuring that the object is initialized with the correct offset
before actually allocating it and appending it into the array.

This commit is based on 074a8b223cb2f419d7a67f83be583360134253ca which
refactors the allocator similarly, but for the ptypes.pstruct library.