Modified the ptype.wrapper_t from the ptypes library to avoid referencing its ._value_ attribute during usage.
Previously, the ptype.wrapper_t class would query its ._value_ attribute
whenever explicitly assigning an object to the backing type. This was
done so that an incompatible type would get casted to the type that was
assigned during its definition. I'm not sure what the original reason
for this was, but a couple of the testcases that were written for the
ptype.wrapper_t class explicitly tested against this functionality.

The issue here is that the original implementation meant that any
initialization of the type (from its definition), and any size
modifications done during runtime (via .alloc) ended up being tightly
coupled. This violates the intention of using the .alloc method for
changing the boundaries throughout all available base types and would
have resulted in the backing type being unable to be re-allocated using
the standard .alloc method. This means that the wrapper_t.alloc method
was only able to allocate using the backing type from the definition,
with the effect that any runtime changes to the type would end up being
lost. This is due to the default generic.alloc method falling back to
using generic.set to apply non-type changes to the allocated type.

This implementation resulted in a silent issue where any assignments to
the .object attribute would have the the effect of logging a debug event
which was never seen during normal usage.

As the previous implementation greatly limits the flexibility of the
ptype.wrapper_t base class, the object that is allocated during regular
usage has been completely decoupled from the object that is specified in
the definition. Thus, assigning an instance to the wrapper_t.object
property will now make a copy of the instance being assigned, re-parent
it, and then assign the re-parented backing type into the correct
property of the instance. After the backing type has been corrected, its
serialized data will be written into the wrapper_t.value property to
ensure that the size is immediately up-to-date. These changes should
allow the wrapper_t and its derivatives to have their backing type
changed or restored on an as-needed basis.

This was accomplished by implementing the wrapper_t.alloc method. The
commit, however, only clears the backing type so that the reallocation
can be applied. Future commits should implement the rest of this method
so that a caller is able to properly change the size of the backing type
or initialize the backing type with a default value as determined by the
implementation of the backing type. The original unit tests that were
depending on the implicit casting done by the wrapper_t.object property
have been removed. Instead, they now test whether the property can be
used to change and restore the backing type after it has been allocated.

These changes introduced in this commit might have an impact on both the
ptype.encoded_t and ptype.pointer_t classes. In my tests, though, all
of the pointer types still seem to work correctly. Despite this, more
unit-tests will definitely need to be written.