Hacked up the osi templates to add a ton of quality-of-life improvements.
Each of the layer submodules now have enumerations that describe the
format of the layer that follows it. This involved tweaking the main
layers class so that an implementor can return an actual abstract type
for the layer identifier from the `stackable.layer` method. The addition
of this identifier allows the newly added `layers.append` method to
automatically set the correct field in the previous layer with the
correct identifier for the new type being appended to the layers.

An alias for the default header (defined as "header") has also been
added to all of the available layer sub-modules. Originally, the
`network.arp` templates contained all of the protocol numbers and
link-layer types. At the moment, these tables have now been moved to the
`layers` type for each submodule as an enumeration for each header that
has been registered. In the future, whenever the layer header
registration gets refactored, these enumerations should be localized
into a single place that is within scope of each of the submodules.