Updated the method names for the `provider.remote` class from the ptypes library.
Originally, the `remote` provider used the implementation of the
`remote.read` method in order to read data from the remote source. Due
to terrible foresight, the semantics of this name doesn't correlate to
its paired method, `remote.send`. To remedy this, the `remote.read`
method was renamed to `remote.receive`. In order to support backwards
compatibility, the default implementation of `remote.receive` will
attempt to dispatch to `remote.read` if it was implemented.