Added a guard that warns when a loaderror is raised by a demoted `ptypes.parray.block` array.
The implementation of the `parray.block` from the ptypes library is
instantiated with a "blocksize" attribute which is used to determine how
many elements to load from a provider. However, the `parray.block` class
has an option to fall-back to the original array loading methodology when an
explicit "length" has been assigned to the instance.

Unfortunately if the demoted array failed to load, a really confusing
exception will be raised. This should only occur when the user has
mistakenly assigned the "length" attribute instead of the blocksize.

To remedy this, we catch the `LoadError` exception from a demoted block
array, log a warning explaining why the load failed, and then re-raise
the exception. Hopefully this provides enough guidance to the user to
know why the load failed, and still output the backtrace that is
produced from an intentional or expected loading error.