Renamed the `image.webp` package template to `image.webp.riff` prior to its refactor.
Earlier, some preliminary definitions were written for the "VP8"
bitstream and added to the `image.webp` template module. As this
bitstream is specific to WEBP only, it was decided that the `image.webp`
templates should be promoted to a template library. As a result, the
commits prior to this one move the original definitions into a package
directory.

This commit accommodates the refactor by moving the original definitions
for the `image.webp` library into the `image.webp.riff` template module
which is more accurately named. This way the `File` structure from the
original definitions and the application can be moved into their own
template or package files.