Fixed a bug in the parray.terminated class from the ptypes library when allocating with a length.
The implementation of the `parray.terminated` class is intended to
act like a `parray.type`, but with the additional ability to add a
method for terminating the decoding of said array. If this method is not
implemented, the parray.terminated class should act like the regular
`parray.type`.

When instantiating a type, the user can explicitly specify attributes to
be assigned directly to the instance. Similar to that capability, the
`alloc()` or `load()` methods of the instance can be used to explicitly
specify attributes to be temporarily applied during the allocation or
loading processes.

The `parray.terminated` had an issue in its implementation of
`parray.terminated.alloc` where the "length" attribute would also be
assigned disregarding any already-existing assignments of the attribute.
This resulted in an instance of the terminated array type not having
the expected length if allocating without any explicit attributes or
fields.

This commit fixes the described issue by modifying the method to alloc
the length giving the priority to the instance attribute when no there
are no fields passed to it. Some unit tests that exercise the
implementation were also included.