Updated the definition for the `Extension` type from the `protocol.tlsext` templates.
The original `Extension` type would simply look up the extension number
and return the matching definition. This works for most purposes, but it
has the issue that if an extension is not decoded properly and returns
an incorrect length, then the entire extension array will be affected by
said length. This commit adds another "padding" field to the `Extension`
type in case the process of decoding the extension does not result in
the length being properly calculated.

The definition for the "data" field from the `Extension` class was also
updated to set the correct length on the extension type being returned
if said extension type belongs to a `ptype.block` class.